refactor(web): pass generic userHasEditPermission to mapper instead of session data
