+++ appeals/web/src/server/lib/mappers/permissions.mapper.js
@@ -6,7 +6,17 @@
  * @returns {ActionItemProperties | undefined}
  */
 export const mapActionComponent = (permission, session, component) => {
-	if (session.permissions && session.permissions[permission]) {
+	if (userHasPermission(permission, session)) {
 		return component;
 	}
+};
+
+/**
+ *
+ * @param {string} permission
+ * @param {import("express-session").Session & Partial<import("express-session").SessionData>} session
+ * @returns {boolean}
+ */
+export const userHasPermission = (permission, session) => {
+	return Boolean(session.permissions && session.permissions[permission]);
 };
